requirements.txt
-i https://pypi.org/simple
asgiref==3.2.7
django-environ==0.4.5
django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
mysqlclient==1.4.6
pytz==2019.3
raven==6.10.0
sqlparse==0.3.1
